update_gen: topology with single BR process (#43)
update_gen now generates a topology configuration with a single Border Router (BR) process with multiple interfaces each.

* each interface is configured to use a separate internal address, as otherwise the IFID packets are not forwarded correctly (fixed in scionproto/scion#1587)
* only a fullsync actually "upgrades" the config to use this new single-BR configuration. Using with deltasync will only use a "shared" BR for newly created or modified interfaces.
* depends on netsec-ethz/netsec-scion#4
* add various unit tests for the syncing logic, some changes to the code structure because of that.

Note: the logic has gotten more complicated as the references into the "InternalAddrs" list have to be maintained explicitly. This can / has to be removed with the breaking changes in scionproto/scion#1587, as support for multiple internal addresses will be removed.